serverless-crud/requirements.txt
pulumi>=3
pulumi_aws>=6.13.3
python-dotenv>=1.0.0